+++ WebBanXeDapDien/src/main/java/app/model/ProductsColorsEntity.java
@@ -5,16 +5,16 @@ import java.io.Serializable;
 import java.util.Objects;
 
 public class ProductsColorsEntity implements Serializable {
-    private short productId ;
+    private Integer productId ;
     private short color;
     private int amount;
     private ProductsEntity productsByProductId;
 
-    public short getProductId() {
+    public Integer getProductId() {
         return productId;
     }
 
-    public void setProductId(short productId) {
+    public void setProductId(Integer productId) {
         this.productId = productId;
     }
